requirements.txt
pandas==2.1.0
numpy==1.24.0
scikit-learn==1.3.0
tensorflow==2.13.0
keras==2.13.1
psutil==5.9.5
streamlit==1.27.0
matplotlib==3.7.2
seaborn==0.12.2
tldextract==3.4.4
pyyaml==6.0.1
joblib==1.3.2
plotly==5.17.0
jupyter==1.0.0
ipykernel==6.25.0